+++ packages/docs/.vitepress/theme/components/VuejsdeConfBanner.vue
@@ -1,11 +1,18 @@
 <template>
   <div class="banner banner-vuejsconf" v-if="isVisible">
-    <a href="https://conf.vuejs.de/?utm_source=vuejs&utm_medium=referral&utm_campaign=banner-placement&utm_content=banner"
+    <a href="https://conf.vuejs.de/tickets/?voucher=COMMUNITY&utm_source=vuejs&utm_medium=referral&utm_campaign=banner-placement&utm_content=banner"
       target="_blank">
       <picture>
-        <source media="(min-width:1200px)" srcset="/vuejsde-conf/vuejsdeconf_banner_large.png" />
-        <source media="(min-width:920px)" srcset="/vuejsde-conf/vuejsdeconf_banner_medium.png" />
-        <img src="/vuejsde-conf/vuejsdeconf_banner_small.png" alt="" />
+        <source media="(min-width:1260px)"
+          srcset="/vuejsde-conf/vuejsdeconf_banner_large.png, /vuejsde-conf/vuejsdeconf_banner_large_2x.png 2x" />
+        <source media="(min-width:970px)"
+          srcset="/vuejsde-conf/vuejsdeconf_banner_medium.png, /vuejsde-conf/vuejsdeconf_banner_medium_2x.png 2x" />
+        <source media="(min-width:576px)"
+          srcset="/vuejsde-conf/vuejsdeconf_banner_small.png, /vuejsde-conf/vuejsdeconf_banner_small_2x.png 2x" />
+        <source media="(min-width:320px)"
+          srcset="/vuejsde-conf/vuejsdeconf_banner_smallest.png, /vuejsde-conf/vuejsdeconf_banner_smallest_2x.png 2x"
+          alt="" />
+        <img src="/vuejsde-conf/vuejsdeconf_banner_smallest_2x.png" alt="" />
       </picture>
     </a>
     <div class="close-btn" @click.stop.prevent="closeBanner">
@@ -19,7 +26,7 @@ import { ref, onMounted } from 'vue'
 
 const isVisible = ref(false)
 
-const nameStorage = 'VUEJSDECONF-BANNER-JUNE-2024'
+const nameStorage = 'VUEJSDECONF-BANNER-SEPTEMBER-2024'
 
 const resetLayoutTopHeight = () => {
   document.documentElement.classList.add('banner-dismissed')
@@ -43,7 +50,7 @@ onMounted(() => {
 </script>
 <style>
 html:not(.banner-dismissed) {
-  --vp-layout-top-height: 60px;
+  --vp-layout-top-height: 72px;
 }
 </style>
 <style scoped>
@@ -59,7 +66,7 @@ html:not(.banner-dismissed) {
   text-align: center;
   font-size: 12px;
   font-weight: 600;
-  color: #fff;
+  color: #000;
 }
 
 .banner-dismissed .banner {
@@ -70,8 +77,12 @@ a {
   text-decoration: underline;
 }
 
+.close {
+  font-size: 24px;
+}
+
 .banner-vuejsconf {
-  background: linear-gradient(90deg, #fff 50%, #43b883 50%);
+  background: linear-gradient(90deg, #fff 50%, #6f97c4 50%);
 }
 
 .banner-vuejsconf a {
@@ -80,8 +91,8 @@ a {
 }
 
 .banner-vuejsconf .close-btn {
-  top: 10px;
-  right: 10px;
+  top: 0px;
+  left: 0px;
   z-index: 99;
   position: absolute;
   border-radius: 50%;